+++ src/argilla_server/apis/v0/handlers/workspaces.py
@@ -21,6 +21,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from argilla_server.contexts import accounts
 from argilla_server.database import get_async_db
 from argilla_server.errors import EntityAlreadyExistsError, EntityNotFoundError
+from argilla_server.errors.future import NotUniqueError
 from argilla_server.policies import WorkspacePolicy, WorkspaceUserPolicy, authorize
 from argilla_server.pydantic_v1 import parse_obj_as
 from argilla_server.schemas.v0.users import User
@@ -39,10 +40,10 @@ async def create_workspace(
 ):
     await authorize(current_user, WorkspacePolicy.create)
 
-    if await accounts.get_workspace_by_name(db, workspace_create.name):
+    try:
+        workspace = await accounts.create_workspace(db, workspace_create.dict())
+    except NotUniqueError:
         raise EntityAlreadyExistsError(name=workspace_create.name, type=Workspace)
-
-    workspace = await accounts.create_workspace(db, workspace_create)
 
     return Workspace.from_orm(workspace)